+++ opentelemetry-appender-log/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+### Fixed
+
+- - Add log appender versions to loggers (#1182).
+
 ## v0.1.0
 
 Initial crate release