+++ opentelemetry-appender-tracing/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## vNext
 
+- Add log appender versions to loggers (#1182)
+
 ## v0.1.0
 
 Initial crate release